[FIX] 레이아웃 수정사항 반영
- 서브메뉴도 메인 레이아웃 파일에서 관리하도록 햇습니다
- 방명록 구조 살짝 수정
