+++ src/components/Board/PostBoard.js
@@ -6,7 +6,7 @@ export default function PostBoardLayout(props){
     return(
         <>
         <div className="w-full">
-                <div className="flex flex-wrap border-l border-t border-black overflow-y-auto">
+                <div className="flex flex-wrap border-l border-t border-black">
                     <>
                         {dataList.map((data) => (
                             <Post id={data.id} key={data.id} from={data.from} to={data.to} text={data.text}></Post>
